GCP_Functions/requirements.txt
google-cloud-bigquery
google-cloud-storage
pandas
nums_from_string